enhance saga pattern

+++ order/src/main/java/vn/whatsenglish/order/entity/Order.java
@@ -30,13 +30,13 @@ public class Order {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "order_code")
     private String orderCode;
 
     @Column(name = "customer_id")
-    private Long customerId;
+    private Integer customerId;
 
     @Column(name = "shipping_address_id")
     private Long shippingAddressId;
